notifications + moved files
