+++ app/src/main/res/layout/cities_item.xml
@@ -1,26 +1,27 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:layout_marginStart="20dp"
-    android:layout_marginEnd="20dp"
     android:layout_marginTop="12dp"
+    android:layout_marginEnd="20dp"
     android:layout_marginBottom="12dp"
     android:orientation="horizontal">
 
     <TextView
         android:id="@+id/rv_city_name"
-        android:background="@color/light_blue"
-        android:padding="6dp"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
+        android:background="@color/light_blue"
         android:ellipsize="end"
         android:fontFamily="@font/poppins_light"
         android:gravity="start"
+        android:padding="6dp"
         android:paddingStart="20dp"
         android:paddingEnd="20dp"
-        android:lines="1"
-        android:text="@string/city_name"
+        android:singleLine="true"
         android:textColor="@color/white"
-        android:textSize="24sp" />
+        android:textSize="24sp"
+        tools:text="City name" />
 </LinearLayout>
